fix(userspace): clippy/rustc/fmt cleanup across all crates — CI gates green (fmt, clippy -D warnings, build)

+++ ade/src/util/benchmark/window.rs
@@ -1,9 +1,9 @@
 #![allow(dead_code)]
 
-use alloc::string::String;
-use crate::util::benchmark::BenchmarkResult;
 use crate::core::desktop::Desktop;
 use crate::core::window::{AppWindow, VisualFlags, WindowState};
+use crate::util::benchmark::BenchmarkResult;
+use alloc::string::String;
 use alloc::vec::Vec;
 use libsarga::io;
 
@@ -14,15 +14,32 @@ pub(crate) fn bench_create_destroy(desktop: &mut Desktop) -> BenchmarkResult {
     let mut ids = Vec::new();
     for i in 0..n {
         let win = AppWindow {
-            x: i as i32 * 10, y: i as i32 * 10, w: 200, h: 150,
-            prev_x: 0, prev_y: 0, prev_w: 200, prev_h: 150,
+            x: i * 10,
+            y: i * 10,
+            w: 200,
+            h: 150,
+            prev_x: 0,
+            prev_y: 0,
+            prev_w: 200,
+            prev_h: 150,
             title: String::from("Bench"),
-            content: Vec::new(), scroll: 0, id: 0, pid: None,
-            focused: true, dragging: false, drag_ox: 0, drag_oy: 0,
-            state: WindowState::Normal, prev_state: WindowState::Normal,
-            flags: VisualFlags::new(), selection: None, anim: None,
-            closing: false, anim_opacity: 0,
-            always_on_top: false, explorer_id: None,
+            content: Vec::new(),
+            scroll: 0,
+            id: 0,
+            pid: None,
+            focused: true,
+            dragging: false,
+            drag_ox: 0,
+            drag_oy: 0,
+            state: WindowState::Normal,
+            prev_state: WindowState::Normal,
+            flags: VisualFlags::new(),
+            selection: None,
+            anim: None,
+            closing: false,
+            anim_opacity: 0,
+            always_on_top: false,
+            explorer_id: None,
         };
         ids.push(desktop.wm.create(win));
     }
@@ -32,7 +49,9 @@ pub(crate) fn bench_create_destroy(desktop: &mut Desktop) -> BenchmarkResult {
 
     let elapsed = desktop.clock_ticks - start;
     io::print_str(&alloc::format!(
-        "[bench] window_create_destroy: {} iterations in {} ticks\n", n, elapsed
+        "[bench] window_create_destroy: {} iterations in {} ticks\n",
+        n,
+        elapsed
     ));
     BenchmarkResult {
         name: "window_create_destroy",
